use relation replace exchange and refund and rebuild move component without batch insert

+++ src/main/java/com/airline/account/model/acca/Iwb.java
@@ -51,7 +51,7 @@ public class Iwb {
 
     private String paxType;
 
-    private double paxQty;
+    private Integer paxQty;
 
     private double luggageHeight;
